requirements.txt
streamlit>=1.35.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
faiss-cpu>=1.8.0
pypdf>=4.2.0
openai>=1.30.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
